rafactoring, moving parts from requestdetails to activepongs and receipt
