+++ screens/ActivePongs.jsx
@@ -1,0 +1,131 @@
+import React from "react"
+import { View, StyleSheet, Text, TouchableHighlight } from "react-native"
+import { useSelector } from "react-redux"
+import { cancelRequest } from "../modules/tripActions"
+import { Icon } from "react-native-elements"
+import Receipt from "./Receipt"
+
+const ActivePongs = () => {
+  const myPong = useSelector((state) => state.myPong)
+  const cancelledRequestResponse = useSelector(
+    (state) => state.cancelledRequestResponse
+  )
+
+  const pongStatus = myPong.status
+
+  let statusColor
+  if (pongStatus === "pending") {
+    statusColor = (
+      <>
+        <Text style={styles.status}>Your request is </Text>
+        <Text style={styles.pending}>{pongStatus}</Text>
+      </>
+    )
+  } else if (pongStatus === "accepted") {
+    statusColor = (
+      <>
+        <Text style={styles.status}>Your request is </Text>
+        <Text style={styles.accepted}>{pongStatus}</Text>
+      </>
+    )
+  } else {
+    statusColor = (
+      <>
+        <Text style={styles.status}>Your request is </Text>
+        <Text style={styles.rejected}>{pongStatus}</Text>
+      </>
+    )
+  }
+
+  if ()
+
+  return (
+    <View style={styles.pong}>
+      <View style={styles.statusContainer}>{statusColor}</View>
+      <View style={styles.itemContainer}>
+        <Icon name="ios-cart" type="ionicon" />
+        <Text style={styles.item}>{myPong.item1}</Text>
+      </View>
+      <View style={styles.itemContainer}>
+        <Icon name="ios-cart" type="ionicon" />
+        <Text style={styles.item}>{myPong.item2}</Text>
+      </View>
+      <View style={styles.itemContainer}>
+        <Icon name="ios-cart" type="ionicon" />
+        <Text style={styles.item}>{myPong.item3}</Text>
+      </View>
+
+      <View style={styles.buttonContainer}>
+        {cancelledRequestResponse ? (
+          <Text id="cancel-message">{cancelledRequestResponse}</Text>
+        ) : (
+          <TouchableHighlight
+            style={styles.cancelButton}
+            onPress={() => {
+              cancelRequest(myPong.id, dispatch)
+            }}
+          >
+            <Text id={"cancel-button"} style={styles.requestButtonText}>
+              Cancel Pong Request
+            </Text>
+          </TouchableHighlight>
+        )}
+      </View>
+      <Receipt />
+    </View>
+  )
+}
+
+const styles = StyleSheet.create({
+  pong: {
+    padding: 10,
+    margin: 10,
+    borderRadius: 5,
+    backgroundColor: "white",
+    shadowColor: "black",
+    shadowOpacity: 2.0,
+  },
+  item: {
+    fontSize: 18,
+    margin: 10,
+  },
+  itemContainer: {
+    flex: 1,
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "left",
+    marginLeft: 15,
+  },
+  buttonContainer: {
+    flex: 1,
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  cancelButton: {
+    height: 30,
+    borderColor: "white",
+    borderWidth: 2,
+    borderRadius: 10,
+    backgroundColor: "#B27183",
+    marginTop: 15,
+    margin: 5,
+    paddingTop: 16,
+    paddingBottom: 18,
+    width: "40%",
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  requestButtonText: {
+    color: "white",
+    fontSize: 12,
+  },
+  statusContainer: {
+    flex: 1,
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+  }
+})
+
+export default ActivePongs
